backend/requirements.txt
Django>=5.2.7
mysqlclient>=2.2.0
PyMySQL>=1.1.0
djangorestframework>=3.15.1
djangorestframework-simplejwt>=5.3.1
django-cors-headers>=4.3.1

# Dependencias internas de Django y DRF
sqlparse>=0.5.0
pytz>=2024.1
tzdata>=2024.1

# Para dependencias de frontend, ver package.json y README.md
